+++ priv/repo/migrations/20181018073059_add_subscriptions_table.exs
@@ -6,11 +6,11 @@ defmodule TdAudit.Repo.Migrations.AddSubscriptionsTable do
 
   def change do
     create table(:subscriptions) do
-      add :event, :string
-      add :resource_id, :integer
-      add :resource_type, :string
-      add :user_email, :string
-      add :periodicity, :string
+      add(:event, :string)
+      add(:resource_id, :integer)
+      add(:resource_type, :string)
+      add(:user_email, :string)
+      add(:periodicity, :string)
 
       timestamps()
     end