+++ priv/repo/migrations/20181121145309_add_default_value_to_last_consumed_event.exs
@@ -3,7 +3,7 @@ defmodule TdAudit.Repo.Migrations.AddDefaultValueToLastConsumedEvent do
 
   def change do
     alter table(:subscriptions) do
-      modify :last_consumed_event, :utc_datetime, default: fragment("now()")
+      modify(:last_consumed_event, :utc_datetime, default: fragment("now()"))
     end
   end
 end