+++ test/support/channel_case.ex
@@ -30,10 +30,11 @@ defmodule TdAuditWeb.ChannelCase do
 
   setup tags do
     :ok = Sandbox.checkout(Repo)
+
     unless tags[:async] do
       Sandbox.mode(Repo, {:shared, self()})
     end
+
     :ok
   end
-
 end